+++ src/components/feature-list/index.js
@@ -1,0 +1,2 @@
+export { default as FeatureListSubscriber } from './FeatureListSubscriber';
+export { default as FeatureList } from './FeatureList';
