+++ src/components/gating/GatingSubscriber.js
@@ -10,7 +10,7 @@ import {
   useDeckCanvasSize, useReady, useUrls, useExpressionValueGetter,
 } from '../hooks';
 import { setObsSelection, mergeObsSets } from '../utils';
-import { getCellSetPolygons } from '../sets/cell-set-utils';
+import { getCellSetPolygons } from '../obs-sets/cell-set-utils';
 import {
   useObsSetsData,
   useFeatureSelection,
