--- src/components/genes/index.js
@@ -1,2 +1,0 @@
-export { default as GenesSubscriber } from './GenesSubscriber';
-export { default as Genes } from './Genes';
