+++ src/components/obs-sets/index.js
@@ -1,0 +1,2 @@
+export { default as ObsSetsManagerSubscriber } from './ObsSetsManagerSubscriber';
+export { default as SetsManager } from './SetsManager';
