--- src/components/sets/index.js
@@ -1,4 +1,0 @@
-export { default as CellSetsManagerSubscriber } from './CellSetsManagerSubscriber';
-export { default as SetsManager } from './SetsManager';
-export { default as CellSetSizesPlot } from './CellSetSizesPlot';
-export { default as CellSetSizesPlotSubscriber } from './CellSetSizesPlotSubscriber';
