+++ src/components/statistical-plots/CellSetExpressionPlot.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import clamp from 'lodash/clamp';
 import { VegaPlot, VEGA_THEMES, DATASET_NAME } from '../vega';
-import { colorArrayToString } from './utils';
+import { colorArrayToString } from '../obs-sets/utils';
 import { capitalize } from '../../utils';
 
 /**
