+++ src/components/statistical-plots/index.js
@@ -1,0 +1,3 @@
+export { default as CellSetSizesPlot } from './CellSetSizesPlot';
+export { default as CellSetExpressionPlot } from './CellSetExpressionPlot';
+export { default as ExpressionHistogram } from './ExpressionHistogram';
